修正点阵变量类型的BUG

+++ arduino-ide-Support/[1.6.x]-hardware(library)/hardware/Microduino/avr/libraries/_01_m_Module_Dot_Matrix/Microduino_MatrixBase.h
@@ -73,7 +73,7 @@ class LedControl : public Print {
     void setCursor(int16_t x, int16_t y);
     virtual size_t write(uint8_t);
     void displayChar(int8_t _row, int8_t _col, char _charIndex);
-    void writeString(int time, char * displayString);
+    void writeString(uint16_t time, char * displayString);
 
     void (*Fun)();
